Mpos Order Update Api integration.

+++ app/src/main/java/com/apollopharmacy/mpospharmacistTest/ui/pbas/pickupsummary/model/OMSOrderForwardResponse.java
@@ -299,7 +299,7 @@ public class OMSOrderForwardResponse implements Serializable {
         private double totalDiscAmount;
         @SerializedName("TotalDiscPct")
         @Expose
-        private Integer totalDiscPct;
+        private double totalDiscPct;
         @SerializedName("TotalRoundedAmount")
         @Expose
         private Integer totalRoundedAmount;
@@ -947,11 +947,11 @@ public class OMSOrderForwardResponse implements Serializable {
             this.totalDiscAmount = totalDiscAmount;
         }
 
-        public Integer getTotalDiscPct() {
+        public double getTotalDiscPct() {
             return totalDiscPct;
         }
 
-        public void setTotalDiscPct(Integer totalDiscPct) {
+        public void setTotalDiscPct(double totalDiscPct) {
             this.totalDiscPct = totalDiscPct;
         }
